third_party/WebKit/LayoutTests/platform/win/http/tests/security/powerfulFeatureRestrictions/geolocation-on-secure-origin-in-insecure-origin-expected.txt

CONSOLE ERROR: line 11: This test can not run without mojo
CONSOLE ERROR: line 11: This test can not run without mojo
CONSOLE ERROR: line 11: This test can not run without mojo
CONSOLE ERROR: line 11: This test can not run without mojo
CONSOLE WARNING: line 17: getCurrentPosition() and watchPosition() no longer work on insecure origins. To use this feature, you should consider switching your application to a secure origin, such as HTTPS. See https://goo.gl/rStTGz for more details.
CONSOLE WARNING: line 17: getCurrentPosition() and watchPosition() no longer work on insecure origins. To use this feature, you should consider switching your application to a secure origin, such as HTTPS. See https://goo.gl/rStTGz for more details.
This is a testharness.js-based test.
PASS getCurrentPosition in iframe 
Harness: the test ran to completion.
